Rename Code_20. Validating emails using regular expression to Code_20. Validating emails using regular expressions
In this we are giving a email as input in order to find out whether it met all the contraints of being a valid one.The pattern will match the constraints such as @ (1 Occurrence) , domain is provided , sub-domain , etc. After matching , if the email is a valid one it will console "Valid" and "Please enter a valid email" if the email provided is not a valid one.
Note : It will not check whether that email address is a valid one or whether that domain or sub domain exists or not. It will only check whether the email provided met all the constraints required for a valid email or not.