Create Code_22. Validating Password using regular expression
As we all must have observed that in order to make the password strong we provide the user with some criteria to be met for creating a password such as :

    Password must contain 1 capital letter [A-Z]
    Password must contain 1 small letter [a-z]
    Password must contain 1 number [0-9]
    Password must contain 1 special character [!,@,#,$,%,^,. . . ] etc
    Length of the password must be within the range [6 to 16]

So we created a regex to validate whether the provided password met all the constraints or not. If the provided password met all the constraints it will console "Valid password" and the snippet will console "invalid password " if any of the above mentioned condition is not met.